requirements.txt
discord.py>=2.3.0
aiosqlite>=0.20.0